Added means of stopping string iteration. Removed mask checks.
At the start of each iteration, all ranks check for the existence
of a file STOP_NOW. If this file exists, then the iteration is
stopped (after calling Lammps and interpolating once more).

Removed the huge number of masked array checks, which were added
during debugging. In the future, the parameterization checks should
perhaps be removed, too.